update questionnaire styling
