+++ soar_9/frontend/app/Homescreen.js
@@ -1,9 +1,10 @@
 import React, {useEffect, useState} from 'react';
-import { View, Text, Pressable, StyleSheet, FlatList, TouchableOpacity} from 'react-native';
+import { View, Text, Pressable, FlatList, TouchableOpacity} from 'react-native';
 import { Divider } from 'react-native-paper'; 
+import styles from "./Styles";
 import { sample } from '../sample-apps.js';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { Link, router } from 'expo-router';
+import { router } from 'expo-router';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 
@@ -39,9 +40,9 @@ const Homescreen = () => {
 
     
     return (
-        <View style={styles.savedAppsContainer}>
+        <View style={styles.container}>
             <Text style={styles.Header}> Home Page </Text>
-            <Divider color='black'/>
+            <Divider/>
 
             <FlatList style={styles.appList}
                 data={sampleData.filter(({ saved }) => saved)}
@@ -53,9 +54,9 @@ const Homescreen = () => {
             />
             <View style={{backgroundColor: '#BCB4D2', width: '100%'}}>
                 <Divider/>
-                <Text style={[styles.words, {color: 'black'}]}>Navigation</Text>
+                <Text style={[styles.words, {color: 'black'}]}>Navigation To:</Text>
                 <Pressable style={styles.button} onPress={() => router.replace("/Questionaire")}>
-                    <Text style={styles.words}>Go To Settings ⚙</Text>
+                    <Text style={styles.words}>Settings</Text>
                 </Pressable>
                 <Pressable style={styles.button} onPress={() => router.replace("/Questionaire")}>
                     <Text style={styles.words}>Contacts</Text>
@@ -70,51 +71,3 @@ const Homescreen = () => {
 
 export default Homescreen;
 
-const styles = StyleSheet.create({
-    savedAppsContainer: {
-      backgroundColor: '#8c80ab',
-      padding: 10,
-      flex: 1,
-    },
-    button: {
-        margin: 10,
-        flexDirection:'row',
-        padding: 10,
-        alignItems: 'center',
-        justifyContent: 'center',
-        backgroundColor: '#28264C',
-        borderColor: 'black',
-        borderWidth: 2,
-        borderRadius: 20, 
-        width: '90%',
-    },
-    Header: {
-        fontSize: 40,
-        fontWeight: 'bold',
-        fontFamily: 'monospace',
-        color: 'black',
-        width: '100%',
-        textAlign: 'center'
-    },
-    item: {
-        flex: 1,
-        margin: 10, 
-        fontSize: 30,
-        color: 'white',
-    },
-    icon: {
-        padding: 5,
-        fontSize: 50,
-        color: 'white',
-        borderColor: 'white',
-        borderWidth: 2,
-        borderRadius: 10,
-    },
-    words: {
-        color: 'white',
-        textAlign: 'center',
-        fontSize: 30,
-        fontWeight: 'bold',
-        fontFamily: 'monospace',
-    }
-    });
