+++ soar_9/frontend/app/Questionaire.js
@@ -1,6 +1,8 @@
 // Import necessary components and hooks
 import React, { useState, useEffect } from 'react';
-import { StyleSheet, View, Text, Switch, TextInput, Button, ScrollView, Pressable, Modal} from 'react-native';
+import { StyleSheet, View, Text, Switch, TextInput, Button, ScrollView, Pressable, Modal, FlatList} from 'react-native';
+import { Divider } from 'react-native-paper'; 
+import styles from "./Styles";
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import Slider from '@react-native-community/slider';
 import {sample} from '../sample-apps';
@@ -85,187 +87,125 @@ const Questionnaire = () => {
   }, [selectedApps]);
 
   return (
-    <ScrollView style={styles.container}>
-        <ScrollView contentContainerStyle={styles.content}>
-      <Text>Accessibility Settings:</Text>
-      <Text>Enable Speech to Text:</Text>
-      <Switch
-        trackColor={{ false: "#767577", true: "#81b0ff" }}
-        thumbColor={speechToTextEnabled ? "#f5dd4b" : "#f4f3f4"}
-        onValueChange={setSpeechToTextEnabled}
-        value={speechToTextEnabled}
-      />
+    <View style={styles.container}>
+      <Text style={styles.Header}> Settings </Text>
+      <Divider/>
 
-    <Text>Font Size:</Text>
-    <View style={styles.fontSizeOptions}>
-    {['Small', 'Medium', 'Large'].map((size) => (
-        <Pressable
-        key={size}
-        onPress={() => setFontSize(size)}
-        style={[styles.fontSizeButton, fontSize === size && styles.selectedButton]}
-        >
-        <Text style={styles.buttonText}>{size}</Text>
-        </Pressable>
-    ))}
-    
-    </View>
-    
+      <ScrollView>
+        <View style={styles.question}>
+          <Text style={styles.questionfont}>Enable Speech to Text:</Text>
+          <Switch
+            // trackColor={{ false: "#767577", true: "#49476b" }}
+            // thumbColor={speechToTextEnabled ? "white" : "#f4f3f4"}
+            onValueChange={setSpeechToTextEnabled}
+            value={speechToTextEnabled}
+          />
+        </View>
+        <Divider/>
 
-      <Text>Primary Language:</Text>
-      <TextInput
-        style={styles.input}
-        onChangeText={setPrimaryLanguage}
-        value={primaryLanguage}
-        placeholder="Enter Primary Language"
-      />
+        <View style={styles.question}>
+          <Text style={styles.questionfont}>Font Size:</Text>
+          <View style={styles.words}>
+            {['Small', 'Medium', 'Large'].map((size) => (
+              <Pressable
+                key={size}
+                onPress={() => setFontSize(size)}
+                style={[styles.button, fontSize === size && styles1.selectedButton]}
+              >
+                <Text style={[styles.words, {fontSize: 20}]}>{size}</Text>
+              </Pressable>
+            ))}
+          </View>
+        </View>
+        <Divider/>
 
-      <Text>Brightness:</Text>
-      <Slider
-        style={styles.slider}
-        minimumValue={5}
-        maximumValue={100}
-        step={1}
-        value={brightness}
-        onValueChange={setBrightness}
-        minimumTrackTintColor="#FFFFFF"
-        maximumTrackTintColor="#000000"
-        thumbTintColor="#007AFF"
-        orientation="vertical" // Only supported on Android
-      />
-      <Text>{`Brightness: ${brightness}%`}</Text>
+        <View style={styles.question}>
+          <Text style={styles.questionfont}>Primary Language:</Text>
+          <TextInput
+            style={styles.input}
+            onChangeText={setPrimaryLanguage}
+            value={primaryLanguage}
+            placeholder="Enter Primary Language"
+          />
+        </View>
+        <Divider/>
 
-      <View style={styles.container}>
-      <Text style={styles.title}>Select Your Favorite Apps:</Text>
-      <Pressable style={styles.button} onPress={() => setModalVisible(true)}>
-        <Text style={styles.buttonText}>Choose Apps</Text>
+        <View style={styles.question}>
+          <Text style={styles.questionfont}>Brightness: {brightness}%</Text>
+          <Slider
+            style={styles.slider}
+            minimumValue={0}
+            maximumValue={100}
+            step={1}
+            value={brightness}
+            onValueChange={setBrightness}
+            minimumTrackTintColor="#FFFFFF"
+            maximumTrackTintColor="#28264C"
+            thumbTintColor="#28264C"
+            orientation="vertical" // Only supported on Android
+          />
+        </View>
+
+        <View style={styles.question}>
+          <Text style={styles.questionfont}>Select Your Favourite Apps:</Text>
+
+          <Pressable style={styles.button} onPress={() => setModalVisible(true)}>
+            <Text style={[styles.words, {fontSize: 20}]}>Choose Apps</Text>
+          </Pressable>
+
+          <Modal
+            animationType="slide"
+            transparent={true}
+            visible={modalVisible}
+            onRequestClose={() => setModalVisible(!modalVisible)}
+          >
+            <View style={styles.modalView}>
+              <ScrollView>
+                {sample.map((app, index) => (
+                  <Pressable
+                    key={index}
+                    style={styles1.appOption}
+                    onPress={() => toggleAppSelection(app.appName)}
+                  >
+                    <Text style={{ color: selectedApps.includes(app.appName) ? '#FF6347' : '#000' }}>
+                    {selectedApps.includes(app.appName) ? app.appName + ' [remove]': app.appName + ' [add]' }
+                    </Text>
+                  </Pressable>
+                ))}
+              </ScrollView>
+              <Pressable
+                style={styles.button}
+                onPress={() => setModalVisible(!modalVisible)}
+              >
+                <Text style={[styles.words, {fontSize:20}]}>Close</Text>
+              </Pressable>
+            </View>
+          </Modal>
+
+          {/* {selectedApps.map((app, index) => (
+            <Text key={index} style={styles1.selectedApp}>{app}</Text>
+          ))} */}
+
+        </View>
+      </ScrollView>
+
+      <Divider/>
+      <Pressable style={styles.button} onPress={savePreferences}>
+        <Text style={[styles.words, {fontSize:20}]}>Go To Home Page</Text>
       </Pressable>
-
-      <Modal
-        animationType="slide"
-        transparent={true}
-        visible={modalVisible}
-        onRequestClose={() => setModalVisible(!modalVisible)}
-      >
-
-      <View style={styles.modalView}>
-        <ScrollView>
-          {sample.map((app, index) => (
-            <Pressable
-              key={index}
-              style={styles.appOption}
-              onPress={() => toggleAppSelection(app.appName)}
-            >
-              <Text style={{ color: selectedApps.includes(app.appName) ? '#FF6347' : '#000' }}>
-              {selectedApps.includes(app.appName) ? app.appName + ' [remove]': app.appName + ' [add]' }
-              </Text>
-            </Pressable>
-          ))}
-        </ScrollView>
-        <Pressable
-          style={styles.buttonClose}
-          onPress={() => setModalVisible(!modalVisible)}
-        >
-          <Text style={styles.textStyle}>Close</Text>
-        </Pressable>
-      </View>
-
-    </Modal>
-
-    
-    {selectedApps.map((app, index) => (
-      <Text key={index} style={styles.selectedApp}>{app}</Text>
-    ))}
-  </View>
-  
-  </ScrollView>
-    <View style={styles.footer}>
-        <Button title="save preferences and go to home" onPress={savePreferences} />
     </View>
-  </ScrollView>
-  
   );
 };
 
 export default Questionnaire;
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    margin: 10,
-    minHeight:'80vh',
-  },
-  content: {
-    // Styles for your content container
-    flexGrow: 1, // Allows the container to grow to fill the space
-    justifyContent: 'center', // Centers content vertically in the container
-    paddingHorizontal: 10, // Adds some horizontal padding
-  },
-  input: {
-    height: 40,
-    margin: 12,
-    borderWidth: 1,
-    padding: 10,
-  },
-  fontSizeOptions: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    margin: 10,
-  },
-  fontSizeButton: {
-    padding: 10,
-    borderWidth: 1,
-    borderColor: '#007bff',
-    borderRadius: 5,
-    backgroundColor: '#ffffff',
-  },
-  selectedButton: {
-    backgroundColor: 'black',
-  },
-  buttonText: {
-    color: '#007bff',
-  },
-  slider: {
-    height: 50,
-    width: '100%',
-  },
-
-  modalView: {
-    margin: 20,
-    backgroundColor: "white",
-    borderRadius: 20,
-    padding: 35,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5
-  },
+const styles1 = StyleSheet.create({
   appOption: {
     marginBottom: 15,
     alignItems: 'center',
   },
-  buttonClose: {
-    backgroundColor: "#2196F3",
-    borderRadius: 20,
-    padding: 10,
-    elevation: 2,
-    marginTop: 15,
+  selectedButton: {
+    backgroundColor: 'black',
   },
-  textStyle: {
-    color: "white",
-    fontWeight: "bold",
-    textAlign: "center"
-  },
-  selectedApp: {
-    marginTop: 10,
-  },
-  footer: {
-    // Styles for the footer area containing the button
-    paddingBottom: 20, // Adds some padding at the bottom
-    paddingHorizontal: 10, // Adds some horizontal padding
-  },
+  
 });
